+++ src/main/java/com/besson/arknights/block/custom/fantastic/SpecimenDataBoard.java
@@ -12,10 +12,10 @@ public class SpecimenDataBoard extends FurnitureHorizontalFacingBlock {
     public SpecimenDataBoard(Settings settings) {
         super(settings);
     }
-    private static final VoxelShape SHAPE_N = Block.createCuboidShape(-16, 0, 15, 32, 32, 16);
-    private static final VoxelShape SHAPE_S = Block.createCuboidShape(-16, 0, 0, 32, 32, 1);
-    private static final VoxelShape SHAPE_W = Block.createCuboidShape(15, 0, -16, 16, 32, 32);
-    private static final VoxelShape SHAPE_E = Block.createCuboidShape(0, 0, -16, 1, 32, 32);
+    private static final VoxelShape SHAPE_N = Block.createCuboidShape(-16, 5, 14, 32, 32, 16);
+    private static final VoxelShape SHAPE_S = Block.createCuboidShape(-16, 5, 0, 32, 32, 2);
+    private static final VoxelShape SHAPE_W = Block.createCuboidShape(14, 5, -16, 16, 32, 32);
+    private static final VoxelShape SHAPE_E = Block.createCuboidShape(0, 5, -16, 2, 32, 32);
 
     @Override
     public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {