+++ src/main/java/com/besson/arknights/block/custom/fantastic/WideGrayCarpet.java
@@ -8,11 +8,11 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.shape.VoxelShape;
 import net.minecraft.world.BlockView;
 
-public class EcoExhibitionCabinet extends FurnitureHorizontalFacingBlock {
-    public EcoExhibitionCabinet(Settings settings) {
+public class WideGrayCarpet extends FurnitureHorizontalFacingBlock {
+    public static final VoxelShape SHAPE = Block.createCuboidShape(-16, 0, -16, 32, 0.45, 32);
+    public WideGrayCarpet(Settings settings) {
         super(settings);
     }
-    private static final VoxelShape SHAPE = Block.createCuboidShape(-8, 0, -8, 24, 32, 24);
 
     @Override
     public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {